Remove cache dependency on authentication
